apptrack-backend/requirements.txt
# Core
Django>=3.2,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
django-environ>=0.11.2
python-dotenv>=1.0.0

# Database
psycopg2-binary>=2.9.9
dj-database-url>=2.0.0

# Authentication
djangorestframework-simplejwt>=5.3.1

# File Storage
django-storages>=1.14.0
boto3>=1.34.0
python-magic>=0.4.27

# Background Tasks
celery>=5.3.6
django-celery-beat==2.5.0
redis>=5.0.1

# Email
django-templated-email>=3.0.0

# API Documentation
drf-yasg>=1.21.7

# Testing
pytest>=8.0.0
pytest-django>=4.7.0
pytest-cov>=4.1.0
factory-boy>=3.3.0

# Code Quality
black>=24.1.0
isort>=5.13.2
flake8>=7.0.0
mypy>=1.7.0
django-stubs>=4.2.7
djangorestframework-stubs>=3.14.1

# Production
gunicorn>=21.2.0
whitenoise>=6.6.0

# Development
ipython>=8.18.0
django-debug-toolbar>=4.2.0
django-extensions>=3.2.3
